não realiza operacao em etapas da obra Caso obra nao exista

+++ src/domain/etapas-da-obra/etapas-da-obra.service.ts
@@ -10,6 +10,12 @@ export class EtapasDaObraService {
   constructor(private readonly etapaObraRepository: EtapasDaObraRepository) {}
 
   async create(dto: CreateEtapasDaObraDto): Promise<EtapasDaObra> {
+     const obraExists = await this.etapaObraRepository.checkObraExists(
+      dto.obraId
+    );
+    if (!obraExists) {
+      throw new NotFoundException(`Obra com ID ${dto.obraId} não encontrada`);
+    }
     return this.etapaObraRepository.create(dto);
   }
 
@@ -25,7 +31,11 @@ export class EtapasDaObraService {
     return etapa;
   }
 
-  async update(id: number, dto: UpdateEtapasDaObraDto): Promise<EtapasDaObra> {
+  async update(id: number, dto: UpdateEtapasDaObraDto, idObra: number): Promise<EtapasDaObra> {
+    const obraExists = await this.etapaObraRepository.checkObraExists(idObra);
+    if (!obraExists) {
+      throw new NotFoundException(`Obra com ID ${idObra} não encontrada`);
+    }
     const [count, [updated]] = await this.etapaObraRepository.update(id, dto);
     if (count === 0) {
       throw new NotFoundException(`Etapa com ID ${id} não encontrada`);
@@ -33,7 +43,11 @@ export class EtapasDaObraService {
     return updated;
   }
 
-  async remove(id: number): Promise<void> {
+  async remove(id: number, idObra): Promise<void> {
+    const obraExists = await this.etapaObraRepository.checkObraExists(idObra);
+    if (!obraExists) {
+      throw new NotFoundException(`Obra com ID ${idObra} não encontrada`);
+    }
     const deleted = await this.etapaObraRepository.remove(id);
     if (deleted === 0) {
       throw new NotFoundException(`Etapa com ID ${id} não encontrada`);
